redoing chunking system, changing world coords to ints

+++ src/main/java/huntingrl/ecs/components/PositionComponent.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 @Setter
 public class PositionComponent implements Component {
     //x and y are in world coords
-    private double x;
-    private double y;
+    private long x;
+    private long y;
 }
