+++ src/main/java/huntingrl/world/WorldPoint.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class WorldPoint {
-    private double x;
-    private double y;
+    private long x;
+    private long y;
     private short elevation;
 }
